requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==3.0.1
attrs==23.1.0
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
contourpy==1.0.6
cycler==0.11.0
discord==2.2.3
discord.py==2.2.3
distlib==0.3.6
Django==4.2.1
django-csp==3.7
django-hosts==5.2
django-jet-reboot==1.3.3
django-subdomains==2.1.0
feedparser==6.0.10
filelock==3.12.0
fonttools==4.38.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
multidict==6.0.4
mysql-connector-python==8.0.32
numpy==1.23.5
packaging==21.3
Pillow==9.3.0
pip-autoremove==0.10.0
platformdirs==3.5.0
protobuf==3.20.3
psycopg2-binary==2.9.6
pynput==1.7.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
requests==2.28.2
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.14
virtualenv==20.23.0
yarl==1.9.2
